refactor: change colocation of frontend extenders
