+++ js/src/forum/extenders/addRequestDropdown.js
@@ -1,7 +1,7 @@
 import { extend } from 'flarum/common/extend';
 import app from 'flarum/forum/app';
 import HeaderSecondary from 'flarum/forum/components/HeaderSecondary';
-import RequestsDropdown from './components/RequestsDropdown';
+import RequestsDropdown from '../components/RequestsDropdown';
 
 export default function () {
   extend(HeaderSecondary.prototype, 'items', function (items) {
