+++ js/src/forum/extenders/addRequestSetting.js
@@ -2,7 +2,7 @@ import app from 'flarum/forum/app';
 import { extend } from 'flarum/common/extend';
 import Button from 'flarum/common/components/Button';
 import SettingsPage from 'flarum/forum/components/SettingsPage';
-import RequestModal from './components/RequestModal';
+import RequestModal from '../components/RequestModal';
 
 export default function () {
   extend(SettingsPage.prototype, 'accountItems', function (items) {
